pyproject.toml
[project]
name = "routing-with-math"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
